Remove GAWK-specific features
List of gawk-specific features:

* switch-case
* arrays of arrays
* BEGINFILE
* ENDFILE

The switch-case can be easily replaced with if-else structure.

`BEGINFILE` and `ENDFILE` are GAWK triggers. They can be replaced with
functions, e.g, `beginfile()` and `endfile()`. You can call
`beginfile()` when `FNR==1`. You can call `endfile()` when `FNR==1` and
`NR>1`. You must also call `endfile()` inside of `END{}`

Arrays of arrays was replaced with fake multidimentinal array using
index concatenation. For, example `arr[1][2]` was replaced with
`arr[1,2]`. Note that for-in loops don't work with `arr[1,2]`; you have
to use a normal for loop using a variable that you stored the array
length. This was the most difficult to do, because I didn't know the
differences very well.

Note: the only feature that remains is `gensub()`, which can be easily
commented out if needed.